Update map tab tooltips styling and fix pin container behavior
Removed the transparent mouse event attribute from `pin_container` to address potential interaction issues. Updated tooltip styling for pins to improve visual consistency and user readability. These changes enhance the user interface and interactive experience.

If the map window gets really large and is bigger than the mapp showing a solid colored background the pins start to shift. Need to put in an offset if the map container is bigger than the image.